+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Next release
+- Feature: Add event listener to handle language change on web in Virtusize widgets
+
 ### 2.12.3
 - Fix: Align mobile SDK size recommendation API requests with web implementation